Incomplete - task : 01. Find the parameters and initial values.

+++ jbox2d-testbed/src/main/java/cs526/controls/PdController.java
@@ -13,13 +13,13 @@ public abstract class PdController {
 	float Ki = 0;
 	public float torque; 
 	
-	public float moveTo(float targetAngle){
+	public float moveTo(double targetAngle){
 		
-		this.targetAngle = targetAngle;
+		this.targetAngle = (float)targetAngle;
 		float P,  D, diffAngle, derivDiffAngle;
 				
 		currentAngle = getCurrentAngle();
-		diffAngle =  targetAngle - currentAngle;
+		diffAngle =  (float)targetAngle - currentAngle;
 		
 		derivDiffAngle = getAngularVelocity();
